+++ test/supabase_flutter_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
-import 'helpers/mock_storage.dart';
 import 'widget_test_stubs.dart';
 
 void main() {
