+++ test/widget_test_stubs.dart
@@ -81,3 +81,24 @@ void mockAppLink() {
   TestDefaultBinaryMessengerBinding.instance?.defaultBinaryMessenger
       .setMockMethodCallHandler(anotherChannel, (message) => null);
 }
+
+class MockAsyncStorage extends GotrueAsyncStorage {
+  static const pkceHiveBoxName = 'supabase.pkce';
+
+  final Map<String, String> _map = {};
+
+  @override
+  Future<String?> getItem({required String key}) async {
+    return _map[key];
+  }
+
+  @override
+  Future<void> removeItem({required String key}) async {
+    _map.remove(key);
+  }
+
+  @override
+  Future<void> setItem({required String key, required String value}) async {
+    _map[key] = value;
+  }
+}
